preciso fazer mais commits, e melhor descritivos
